requirements.txt

chromadb==1.3.5
dotenv==0.9.9
faiss-cpu==1.13.1
fastapi==0.123.9
gcloud==0.18.3
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-genai==1.54.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gradio==6.1.0
gradio_client==2.0.1
huggingface-hub==0.36.0
ipywidgets==8.1.8
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.1.2
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-google-genai==4.0.0
langchain-huggingface==1.1.0
langchain-text-splitters==1.0.0

multidict==6.7.0
numpy==2.3.5
oauth2client==4.1.3
oauthlib==3.3.1
onnxruntime==1.23.2
pillow==12.0.0

prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20

regex==2025.11.3
requests==2.32.5

safetensors==0.7.0
scikit-learn==1.8.0
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
starlette==0.50.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
uvicorn==0.38.0
